Added reminder validation test first cut. Moved FTs to individual files
